backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.1
attrs==26.1.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.3.0
build==1.4.2
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
deprecation==2.1.0
dnspython==2.8.0
docopt==0.6.2
ecdsa==0.19.2
email-validator==2.3.0
environs==14.6.0
executing==2.2.1
fastapi==0.135.2
fastapi_cors==0.0.6
fastjsonschema==2.21.2
fsspec==2026.3.0
gotrue==2.12.4
greenlet==3.3.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
limits==5.8.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.2.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.1
multidict==6.7.1
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
passlib==1.7.4
pickleshare==0.7.5
pip-tools==7.5.3
platformdirs==4.9.4
postgrest==2.28.3
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyproject_hooks==1.2.0
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyzmq==27.1.0
realtime==2.28.3
referencing==0.37.0
requests==2.33.0
rich==14.3.3
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
supafunc==0.10.2
tenacity==9.1.4
tinycss2==1.4.0
tornado==6.5.5
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
webencodings==0.5.1
websockets==15.0.1
wrapt==2.1.2
yarg==0.1.9
yarl==1.23.0
zstandard==0.25.0
